test 3 coherence salete d unsigned

+++ programmation/sumobot/Table.cpp
@@ -132,13 +132,13 @@ if (donneesCapteurs.avant < 630) // cohérence capteur avant
 {
 	if (donneesCapteurs.avantDroit < 255)
 	{
-		if (donneesCapteurs.avantDroit - donneesCapteurs.avant > MARGE_COHERENCE)
+		if ((int32_t)donneesCapteurs.avantDroit - (int32_t)donneesCapteurs.avant > MARGE_COHERENCE)
 		{
 			donneesCapteurs.avantDroit = 255;
 		}
 		else
 		{
-			if (donneesCapteurs.avantDroit - donneesCapteurs.avant < -MARGE_COHERENCE)
+			if ((int32_t)donneesCapteurs.avantDroit - (int32_t)donneesCapteurs.avant < -MARGE_COHERENCE)
 			{
 				donneesCapteurs.avant = 630;
 			}
@@ -146,13 +146,13 @@ if (donneesCapteurs.avant < 630) // cohérence capteur avant
 	}
 	if (donneesCapteurs.avantGauche < 255)
 	{
-		if (donneesCapteurs.avantGauche - donneesCapteurs.avant > MARGE_COHERENCE)
+		if ((int32_t)donneesCapteurs.avantGauche - (int32_t)donneesCapteurs.avant > MARGE_COHERENCE)
 		{
 			donneesCapteurs.avantGauche = 255;
 		}
 		else
 		{
-			if (donneesCapteurs.avantGauche - donneesCapteurs.avant < -MARGE_COHERENCE)
+			if ((int32_t)donneesCapteurs.avantGauche - (int32_t)donneesCapteurs.avant < -MARGE_COHERENCE)
 			{
 				donneesCapteurs.avant = 630;
 			}
@@ -160,7 +160,7 @@ if (donneesCapteurs.avant < 630) // cohérence capteur avant
 	}
 	if (donneesCapteurs.droit < 255)
 	{
-		if (donneesCapteurs.droit - donneesCapteurs.avant > 0)
+		if ((int32_t)donneesCapteurs.droit - (int32_t)donneesCapteurs.avant > 0)
 		{
 			donneesCapteurs.droit = 255;
 		}
@@ -171,7 +171,7 @@ if (donneesCapteurs.avant < 630) // cohérence capteur avant
 	}
 	if (donneesCapteurs.arriereDroit < 255)
 	{
-		if (donneesCapteurs.arriereDroit - donneesCapteurs.avant > 0)
+		if ((int32_t)donneesCapteurs.arriereDroit - (int32_t)donneesCapteurs.avant > 0)
 		{
 			donneesCapteurs.arriereDroit = 255;
 		}
@@ -182,7 +182,7 @@ if (donneesCapteurs.avant < 630) // cohérence capteur avant
 	}
 	if (donneesCapteurs.arriere < 630)
 	{
-		if (donneesCapteurs.arriere - donneesCapteurs.avant > 0)
+		if ((int32_t)donneesCapteurs.arriere - (int32_t)donneesCapteurs.avant > 0)
 		{
 			donneesCapteurs.arriere = 630;
 		}
@@ -193,7 +193,7 @@ if (donneesCapteurs.avant < 630) // cohérence capteur avant
 	}
 	if (donneesCapteurs.arriereGauche < 255)
 	{
-		if (donneesCapteurs.arriereGauche - donneesCapteurs.avant > 0)
+		if ((int32_t)donneesCapteurs.arriereGauche - (int32_t)donneesCapteurs.avant > 0)
 		{
 			donneesCapteurs.arriereGauche = 255;
 		}
@@ -204,7 +204,7 @@ if (donneesCapteurs.avant < 630) // cohérence capteur avant
 	}
 	if (donneesCapteurs.gauche < 255)
 	{
-		if (donneesCapteurs.gauche - donneesCapteurs.avant > 0)
+		if ((int32_t)donneesCapteurs.gauche - (int32_t)donneesCapteurs.avant > 0)
 		{
 			donneesCapteurs.gauche = 255;
 		}
@@ -219,13 +219,13 @@ if (donneesCapteurs.avantDroit < 255) // coherence capteur avant droit
 {
 	if (donneesCapteurs.avantGauche < 255)
 	{
-		if (donneesCapteurs.avantGauche - donneesCapteurs.avantDroit > MARGE_COHERENCE)
+		if ((int32_t)donneesCapteurs.avantGauche - (int32_t)donneesCapteurs.avantDroit > MARGE_COHERENCE)
 		{
 			donneesCapteurs.avantGauche = 255;
 		}
 		else
 		{
-			if (donneesCapteurs.avantGauche - donneesCapteurs.avantDroit < -MARGE_COHERENCE)
+			if ((int32_t)donneesCapteurs.avantGauche - (int32_t)donneesCapteurs.avantDroit < -MARGE_COHERENCE)
 			{
 				donneesCapteurs.avantDroit = 255;
 			}
@@ -233,13 +233,13 @@ if (donneesCapteurs.avantDroit < 255) // coherence capteur avant droit
 	}
 	if (donneesCapteurs.droit < 255)
 	{
-		if (donneesCapteurs.droit - donneesCapteurs.avantDroit > MARGE_COHERENCE)
+		if ((int32_t)donneesCapteurs.droit - (int32_t)donneesCapteurs.avantDroit > MARGE_COHERENCE)
 		{
 			donneesCapteurs.droit = 255;
 		}
 		else
 		{
-			if (donneesCapteurs.droit - donneesCapteurs.avantDroit < -MARGE_COHERENCE)
+			if ((int32_t)donneesCapteurs.droit - (int32_t)donneesCapteurs.avantDroit < -MARGE_COHERENCE)
 			{
 				donneesCapteurs.avantDroit = 255;
 			}
@@ -247,7 +247,7 @@ if (donneesCapteurs.avantDroit < 255) // coherence capteur avant droit
 	}
 	if (donneesCapteurs.arriereDroit < 255)
 	{
-		if (donneesCapteurs.arriereDroit - donneesCapteurs.avantDroit > 0)
+		if ((int32_t)donneesCapteurs.arriereDroit - (int32_t)donneesCapteurs.avantDroit > 0)
 		{
 			donneesCapteurs.arriereDroit = 255;
 		}
@@ -258,7 +258,7 @@ if (donneesCapteurs.avantDroit < 255) // coherence capteur avant droit
 	}
 	if (donneesCapteurs.arriere < 630)
 	{
-		if (donneesCapteurs.arriere - donneesCapteurs.avantDroit > 0)
+		if ((int32_t)donneesCapteurs.arriere - (int32_t)donneesCapteurs.avantDroit > 0)
 		{
 			donneesCapteurs.arriere = 630;
 		}
@@ -269,7 +269,7 @@ if (donneesCapteurs.avantDroit < 255) // coherence capteur avant droit
 	}
 	if (donneesCapteurs.arriereGauche < 255)
 	{
-		if (donneesCapteurs.arriereGauche - donneesCapteurs.avantDroit > 0)
+		if ((int32_t)donneesCapteurs.arriereGauche - (int32_t)donneesCapteurs.avantDroit > 0)
 		{
 			donneesCapteurs.arriereGauche = 255;
 		}
@@ -280,7 +280,7 @@ if (donneesCapteurs.avantDroit < 255) // coherence capteur avant droit
 	}
 	if (donneesCapteurs.gauche < 255)
 	{
-		if (donneesCapteurs.gauche - donneesCapteurs.avantDroit > 0)
+		if ((int32_t)donneesCapteurs.gauche - (int32_t)donneesCapteurs.avantDroit > 0)
 		{
 			donneesCapteurs.gauche = 255;
 		}
@@ -295,7 +295,7 @@ if (donneesCapteurs.droit < 255) // coherence capteur droit
 {
 	if (donneesCapteurs.avantGauche < 255)
 	{
-		if (donneesCapteurs.avantGauche - donneesCapteurs.droit > 0)
+		if ((int32_t)donneesCapteurs.avantGauche - (int32_t)donneesCapteurs.droit > 0)
 		{
 			donneesCapteurs.avantGauche = 255;
 		}
@@ -306,13 +306,13 @@ if (donneesCapteurs.droit < 255) // coherence capteur droit
 	}
 	if (donneesCapteurs.arriereDroit < 255)
 	{
-		if (donneesCapteurs.arriereDroit - donneesCapteurs.droit > MARGE_COHERENCE)
+		if ((int32_t)donneesCapteurs.arriereDroit - (int32_t)donneesCapteurs.droit > MARGE_COHERENCE)
 		{
 			donneesCapteurs.arriereDroit = 255;
 		}
 		else
 		{
-			if (donneesCapteurs.arriereDroit - donneesCapteurs.droit < -MARGE_COHERENCE)
+			if ((int32_t)donneesCapteurs.arriereDroit - (int32_t)donneesCapteurs.droit < -MARGE_COHERENCE)
 			{
 				donneesCapteurs.droit = 255;
 			}
@@ -320,7 +320,7 @@ if (donneesCapteurs.droit < 255) // coherence capteur droit
 	}
 	if (donneesCapteurs.arriere < 630)
 	{
-		if (donneesCapteurs.arriere - donneesCapteurs.droit > 0)
+		if ((int32_t)donneesCapteurs.arriere - (int32_t)donneesCapteurs.droit > 0)
 		{
 			donneesCapteurs.arriere = 630;
 		}
@@ -331,7 +331,7 @@ if (donneesCapteurs.droit < 255) // coherence capteur droit
 	}
 	if (donneesCapteurs.arriereGauche < 255)
 	{
-		if (donneesCapteurs.arriereGauche - donneesCapteurs.droit > 0)
+		if ((int32_t)donneesCapteurs.arriereGauche - (int32_t)donneesCapteurs.droit > 0)
 		{
 			donneesCapteurs.arriereGauche = 255;
 		}
@@ -342,7 +342,7 @@ if (donneesCapteurs.droit < 255) // coherence capteur droit
 	}
 	if (donneesCapteurs.gauche < 255)
 	{
-		if (donneesCapteurs.gauche - donneesCapteurs.droit > 0)
+		if ((int32_t)donneesCapteurs.gauche - (int32_t)donneesCapteurs.droit > 0)
 		{
 			donneesCapteurs.gauche = 255;
 		}
@@ -357,7 +357,7 @@ if (donneesCapteurs.arriereDroit < 255) // coherence capteur arriere droit
 {
 	if (donneesCapteurs.avantGauche < 255)
 	{
-		if (donneesCapteurs.avantGauche - donneesCapteurs.arriereDroit > 0)
+		if ((int32_t)donneesCapteurs.avantGauche - (int32_t)donneesCapteurs.arriereDroit > 0)
 		{
 			donneesCapteurs.avantGauche = 255;
 		}
@@ -368,13 +368,13 @@ if (donneesCapteurs.arriereDroit < 255) // coherence capteur arriere droit
 	}
 	if (donneesCapteurs.arriere < 630)
 	{
-		if (donneesCapteurs.arriere - donneesCapteurs.arriereDroit > MARGE_AVANT)
+		if ((int32_t)donneesCapteurs.arriere - (int32_t)donneesCapteurs.arriereDroit > MARGE_AVANT)
 		{
 			donneesCapteurs.arriere = 630;
 		}
 		else
 		{
-			if (donneesCapteurs.arriere - donneesCapteurs.arriereDroit < -MARGE_AVANT)
+			if ((int32_t)donneesCapteurs.arriere - (int32_t)donneesCapteurs.arriereDroit < -MARGE_AVANT)
 			{
 				donneesCapteurs.arriereDroit = 255;
 			}
@@ -382,13 +382,13 @@ if (donneesCapteurs.arriereDroit < 255) // coherence capteur arriere droit
 	}
 	if (donneesCapteurs.arriereGauche < 255)
 	{
-		if (donneesCapteurs.arriereGauche - donneesCapteurs.arriereDroit > MARGE_COHERENCE)
+		if ((int32_t)donneesCapteurs.arriereGauche - (int32_t)donneesCapteurs.arriereDroit > MARGE_COHERENCE)
 		{
 			donneesCapteurs.arriereGauche = 255;
 		}
 		else
 		{
-			if (donneesCapteurs.arriereGauche - donneesCapteurs.arriereDroit < -MARGE_COHERENCE)
+			if ((int32_t)donneesCapteurs.arriereGauche - (int32_t)donneesCapteurs.arriereDroit < -MARGE_COHERENCE)
 			{
 				donneesCapteurs.arriereDroit = 255;
 			}
@@ -396,7 +396,7 @@ if (donneesCapteurs.arriereDroit < 255) // coherence capteur arriere droit
 	}
 	if (donneesCapteurs.gauche < 255)
 	{
-		if (donneesCapteurs.gauche - donneesCapteurs.arriereDroit > 0)
+		if ((int32_t)donneesCapteurs.gauche - (int32_t)donneesCapteurs.arriereDroit > 0)
 		{
 			donneesCapteurs.gauche = 255;
 		}
@@ -411,7 +411,7 @@ if (donneesCapteurs.arriere < 630) // coherence capteur arriere
 {
 	if (donneesCapteurs.avantGauche < 255)
 	{
-		if (donneesCapteurs.avantGauche - donneesCapteurs.arriere > 0)
+		if ((int32_t)donneesCapteurs.avantGauche - (int32_t)donneesCapteurs.arriere > 0)
 		{
 			donneesCapteurs.avantGauche = 255;
 		}
@@ -422,13 +422,13 @@ if (donneesCapteurs.arriere < 630) // coherence capteur arriere
 	}
 	if (donneesCapteurs.arriereGauche < 255)
 	{
-		if (donneesCapteurs.arriereGauche - donneesCapteurs.arriere > MARGE_COHERENCE)
+		if ((int32_t)donneesCapteurs.arriereGauche - (int32_t)donneesCapteurs.arriere > MARGE_COHERENCE)
 		{
 			donneesCapteurs.arriereGauche = 255;
 		}
 		else
 		{
-			if (donneesCapteurs.arriereGauche - donneesCapteurs.arriere < -MARGE_COHERENCE)
+			if ((int32_t)donneesCapteurs.arriereGauche - (int32_t)donneesCapteurs.arriere < -MARGE_COHERENCE)
 			{
 				donneesCapteurs.arriere = 630;
 			}
@@ -436,7 +436,7 @@ if (donneesCapteurs.arriere < 630) // coherence capteur arriere
 	}
 	if (donneesCapteurs.gauche < 255)
 	{
-		if (donneesCapteurs.gauche - donneesCapteurs.arriere > 0)
+		if ((int32_t)donneesCapteurs.gauche - (int32_t)donneesCapteurs.arriere > 0)
 		{
 			donneesCapteurs.gauche = 255;
 		}
@@ -451,7 +451,7 @@ if (donneesCapteurs.arriereGauche < 255) // coherence capteur arriere gauche
 {
 	if (donneesCapteurs.avantGauche < 255)
 	{
-		if (donneesCapteurs.avantGauche - donneesCapteurs.arriereGauche > 0)
+		if ((int32_t)donneesCapteurs.avantGauche - (int32_t)donneesCapteurs.arriereGauche > 0)
 		{
 			donneesCapteurs.avantGauche = 255;
 		}
@@ -462,13 +462,13 @@ if (donneesCapteurs.arriereGauche < 255) // coherence capteur arriere gauche
 	}
 	if (donneesCapteurs.gauche < 255)
 	{
-		if (donneesCapteurs.gauche - donneesCapteurs.arriereGauche > MARGE_COHERENCE)
+		if ((int32_t)donneesCapteurs.gauche - (int32_t)donneesCapteurs.arriereGauche > MARGE_COHERENCE)
 		{
 			donneesCapteurs.gauche = 255;
 		}
 		else
 		{
-			if (donneesCapteurs.gauche - donneesCapteurs.arriereGauche < -MARGE_COHERENCE)
+			if ((int32_t)donneesCapteurs.gauche - (int32_t)donneesCapteurs.arriereGauche < -MARGE_COHERENCE)
 			{
 				donneesCapteurs.arriereGauche = 255;
 			}
@@ -479,13 +479,13 @@ if (donneesCapteurs.gauche < 255) // coherence capteur gauche
 {
 	if (donneesCapteurs.avantGauche < 255)
 	{
-		if (donneesCapteurs.avantGauche - donneesCapteurs.arriereGauche > MARGE_COHERENCE)
+		if ((int32_t)donneesCapteurs.avantGauche - (int32_t)donneesCapteurs.arriereGauche > MARGE_COHERENCE)
 		{
 			donneesCapteurs.avantGauche = 255;
 		}
 		else
 		{
-			if (donneesCapteurs.avantGauche - donneesCapteurs.arriereGauche > MARGE_COHERENCE)
+			if ((int32_t)donneesCapteurs.avantGauche - (int32_t)donneesCapteurs.arriereGauche > MARGE_COHERENCE)
 			{
 				donneesCapteurs.arriereGauche = 255;
 			}
